Subdivide co.rsk.rpc.netty package
