+++ rskj-core/src/main/java/co/rsk/RskContext.java
@@ -64,7 +64,11 @@ import co.rsk.rpc.modules.rsk.RskModule;
 import co.rsk.rpc.modules.rsk.RskModuleImpl;
 import co.rsk.rpc.modules.txpool.TxPoolModule;
 import co.rsk.rpc.modules.txpool.TxPoolModuleImpl;
-import co.rsk.rpc.netty.*;
+import co.rsk.rpc.netty.JsonRpcWeb3ServerHandler;
+import co.rsk.rpc.netty.RskJsonRpcHandler;
+import co.rsk.rpc.netty.http.JsonRpcWeb3FilterHandler;
+import co.rsk.rpc.netty.http.Web3HttpServer;
+import co.rsk.rpc.netty.ws.Web3WebSocketServer;
 import co.rsk.scoring.PeerScoring;
 import co.rsk.scoring.PeerScoringManager;
 import co.rsk.scoring.PunishmentParameters;
