+++ rskj-core/src/main/java/co/rsk/rpc/netty/http/Web3ResultHttpResponseHandler.java
@@ -15,8 +15,9 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package co.rsk.rpc.netty;
+package co.rsk.rpc.netty.http;
 
+import co.rsk.rpc.netty.Web3Result;
 import com.googlecode.jsonrpc4j.DefaultHttpStatusCodeProvider;
 import io.netty.channel.ChannelFutureListener;
 import io.netty.channel.ChannelHandlerContext;
